refactor(structured-logs): extract shared types rendering and common helpers
